requirements.txt
tensorflow
gym[all]
numpy
pandas
playwright
requests
beautifulsoup4
celery
mysql-connector-python
